+++ gcs/CHANGES.md
@@ -1,6 +1,7 @@
 # Release Notes
 
 ## Next
+1. Track Client Feature Usage using Custom Header
 
 ## 4.0.1 - 2026-02-09
 1. Add fast fail feature for gRPC.